requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.25.0

# Image processing
Pillow>=8.0.0
numpy>=1.20.0

# Optional: For advanced image manipulation
# opencv-python>=4.5.0
# matplotlib>=3.3.0
